requirements.txt
Flask
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
streamlit>=1.30.0
google-generativeai>=0.8.5
python-dotenv>=1.0.0
gunicorn

